Remove restart_service on service_limits define
Since 97dd16fa32886b5b0f77a6f38a4953d4c1511c0e this parameter is a bad
idea. It doesn't do a daemon reload and it may restart at the wrong
time. In fact, I'd argue it's always been a bad idea.

The recommended alternative to this is to manage the service explicitly
and let Puppet handle it. There is an automatic notify that takes care
of it.

Fixes #190