Add a socket_service definition
For systemd socket activated services it's needed to have both the
.socket and .service unit defintions in place when starting the service.
Prior to 97dd16fa32886b5b0f77a6f38a4953d4c1511c0e the systemctl
daemon-reload took care of it, but now this must be done explicitly.

This new defined type makes it easy and reduces the burden on modules
that define the unit/service pair.